Hello from Jeep Sheep TV
Hey this is Liam from Jeep Sheep TV
I have never used this site so hopefully I'm not messing it up. 
I would love to help you get a 2.5L auto into this database.
It would also be great to adapt the PCM to control the Supercharger.
If possible we could try to use the programming designed for the AC compressor to create a custom tune when the SC is enabled. 
I've never dug into a PCM so I'm not sure if I know what I'm taking about but I'm willing to learn!